split result reader from result helper
