+++ c/driver/postgresql/CMakeLists.txt
@@ -33,6 +33,7 @@ add_arrow_lib(adbc_driver_postgresql
               database.cc
               postgresql.cc
               result_helper.cc
+              result_reader.cc
               statement.cc
               OUTPUTS
               ADBC_LIBRARIES
